refactor: move utils and services into app directory structure
- Move utils/presentation_reader.py to app/utils/presentation_reader.py
- Move utils/gemini_service.py to app/services/gemini_service.py
- Update imports in controllers and services to reflect new paths
- Update unit tests to import from new locations
- Add __init__.py files to define python packages
- Ensure all tests pass
