+++ app/services/gemini_service.py
