+++ app/utils/presentation_reader.py
