--- utils/__init__.py
@@ -1,2 +1,0 @@
-# Utils package for CreateVideo-Website
-
